Adding the updates to the backend-api and frontend
